pyproject.toml
[project]
name = "esmevals"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "biopython>=1.85",
    "fair-esm>=2.0.0",
    "plm-fid>=1.0.0",
    "rich>=14.3.3",
    "torch>=2.8.0",
    "tqdm>=4.67.3",
    "transformers>=4.57.6",
    "typer>=0.24.1",
]

[project.scripts]
esmevals = "ESMEvals.cli:app"
